// src/types/logging/TableOptions.ts
export enum TableOptions {
    ARTICLE = "Article",
    AGENCY = "Agency",
    SECTOR = "Sector",
    STATUS = "Status",
    FIND_AND_REPLACE = "FindAndReplace",
    TAG = "Tag",
    CLASSIFICATION = "Classification",
    FEDERAL_REGISTER = "FederalRegister",
    FILING_TYPE = "FilingType",
    ALERT_TYPE = "AlertType"
}